make useDayFeed and useUserFeed hook
